// src/app/core/interfaces/UserMenuTree.ts
export interface UserMenuTree {
  menuId: number;

  clntId: number;

  usrKey: number;

  description: string;

  pmnDescription: string;

  parentMenuId: number;

  modPageParam: string;

  modPageParam2: string;

  modDisplaySeq: number;

  modMeunType: number;

  modViewName: string;

  modDisplayed: string;

  modPagePath: string;

  modLogoClass: string;

  pmnType: string;

  mlCaption: string;
  langKey: number;
  langShortName: string;
  fcpKey: number;
}