// src/app/core/interfaces/defaultValues.ts
export interface DefaultValues {
  flexField1: string;
  flexField2: string;
  urDownPaymentPercentage: string;
  urFinancePurpose: string;
  urFinanceType: string;
  urGraceDays: string;
  urMasterFinanceConcept: string;
  urPrincipal: string;
  urProfitRate: string;
  urSector: string;
  urTenureNo: string;
  urUtilization: string;
  wcrKey: string;
}