// src/app/core/interfaces/financeProduct.ts
export interface FinanceProduct {
  fcpKey: number;
  productName: string;
  productIconPath: string;
  productTermsConditions: string;
  productRequiredDocuments: string;
  fcpItemsStatus: string;
  langShortName: string;
  langKey: number;
  productDescription: string;
  fcpDownPaymentAllowed: string;
}